temp via objecten

+++ Code/sim-knx/sim-knx/inc/Response.h
@@ -25,7 +25,7 @@ public:
 
     void parse(string);
     string raw(void);
-    int data(void);
+    float data(void);
     bool mode(void);
 
     bool isEmpty(void);
